backend/requirements.txt
Flask==3.1.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.1
Flask-Migrate==4.1.0
psycopg[binary]>=3.2.10
pymongo==4.12.1
openpyxl==3.1.5
python-dotenv==1.1.0
marshmallow==3.26.1
